backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.29.0
sqlalchemy==2.0.41
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.11.4
pydantic-settings==2.8.1
psycopg[binary]==3.2.9
cryptography==42.0.5